--- src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/action/BrAPIState.java
@@ -1,4 +1,0 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.action;
-
-public interface BrAPIState {
-}
