+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/middleware/validator/field/FieldValidator.java
@@ -1,7 +1,8 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.validator;
+package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.validator.field;
 
 import io.micronaut.context.annotation.Primary;
 import org.breedinginsight.api.model.v1.response.ValidationError;
+import org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.validator.field.ObservationValidator;
 import org.breedinginsight.model.Trait;
 
 import javax.inject.Singleton;
