+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/middleware/validator/field/NumericalValidator.java
@@ -1,25 +1,27 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.validator;
+package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.validator.field;
 
 import io.micronaut.http.HttpStatus;
 import lombok.extern.slf4j.Slf4j;
 import org.breedinginsight.api.model.v1.response.ValidationError;
+import org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.validator.field.ObservationValidator;
 import org.breedinginsight.brapps.importer.services.processors.experiment.service.ObservationService;
 import org.breedinginsight.model.Trait;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import static org.breedinginsight.brapps.importer.services.processors.experiment.model.ExpImportProcessConstants.TIMESTAMP_PREFIX;
-import static org.breedinginsight.dao.db.enums.DataType.ORDINAL;
+import static org.breedinginsight.dao.db.enums.DataType.NUMERICAL;
 
 @Slf4j
 @Singleton
-public class OrdinalValidator implements ObservationValidator {
+public class NumericalValidator implements ObservationValidator {
     @Inject
     ObservationService observationService;
 
-    public OrdinalValidator(ObservationService observationService) {
+    public NumericalValidator(ObservationService observationService) {
         this.observationService = observationService;
     }
     @Override
@@ -44,16 +46,22 @@ public class OrdinalValidator implements ObservationValidator {
             return Optional.empty();
         }
 
-        // Skip if this is not an ordinal trait
-        if (!ORDINAL.equals(variable.getScale().getDataType())) {
+        // Skip if this is not a numerical trait
+        if (!NUMERICAL.equals(variable.getScale().getDataType())) {
             return Optional.empty();
         }
 
-        // Validate categories
-        if (!observationService.validCategory(variable.getScale().getCategories(), value)) {
-            return Optional.of(new ValidationError(fieldName, "Undefined ordinal category detected", HttpStatus.UNPROCESSABLE_ENTITY));
-        }
+        Optional<BigDecimal> number = observationService.validNumericValue(value);
+        Optional<ValidationError> validationError = number
+                .flatMap(num -> {
+                    if (observationService.validNumericRange(num, variable.getScale())) {
+                        return Optional.empty(); // Return empty Optional if value is in numeric range
+                    } else {
+                        return Optional.of(new ValidationError(fieldName, "Value outside of min/max range detected", HttpStatus.UNPROCESSABLE_ENTITY));
+                    }
+                });
 
-        return Optional.empty();
+        return validationError;
+
     }
 }
